テスト用の fluent-bit.yml のテンプレートをファイルに切り出す
