+++ ingester/tests/fixtures/fluent-bit/default-fluent-bit.yml.j2
@@ -1,0 +1,50 @@
+env:
+  S3_ENDPOINT: {{ s3_endpoint }}
+  S3_BUCKET: {{ s3_bucket }}
+  S3_PREFIX: {{ s3_prefix }}
+  SORA_LOG_PATH: {{ sora_log_path }}
+
+service:
+  flush:        1
+  daemon:       Off
+  log_level:    info
+  http_server: Off
+
+pipeline:
+  inputs:
+    - name: tail
+      path: ${SORA_LOG_PATH}/rtc_stats.jsonl
+      parser: json
+      read_from_head: true
+      db: /state/rtc_stats.db
+      tag: rtc_stats
+    - name: tail
+      path: ${SORA_LOG_PATH}/session_webhook.jsonl
+      parser: json
+      read_from_head: true
+      db: /state/session_webhook.db
+      tag: session_webhook
+  outputs:
+    - name: s3
+      match: 'rtc_stats'
+      bucket: ${S3_BUCKET}
+      endpoint: ${S3_ENDPOINT}
+      compression: gzip
+      s3_key_format: /${S3_PREFIX}/$TAG/%Y/%m/%d/$UUID.gz
+      upload_timeout: 10s
+      json_date_key: off
+    - name: s3
+      match: 'session_webhook'
+      bucket: ${S3_BUCKET}
+      endpoint: ${S3_ENDPOINT}
+      compression: gzip
+      s3_key_format: /${S3_PREFIX}/$TAG/%Y/%m/%d/$UUID.gz
+      upload_timeout: 10s
+      json_date_key: off
+
+parsers:
+  - name: json
+    format: json
+    time_key: timestamp
+    time_format: '%Y-%m-%dT%H:%M:%S.%L%z'
+    time_keep: on
